+++ src/main/java/auth/dto/MemberDetails.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.dto;
 
 public class MemberDetails {
     private final Long id;
