+++ src/main/java/nextstep/WebMvcConfiguration.java
@@ -1,6 +1,10 @@
 package nextstep;
 
 import auth.*;
+import auth.interceptor.AdminInterceptor;
+import auth.support.LoginMemberArgumentResolver;
+import auth.token.JwtTokenProvider;
+import auth.service.LoginService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
