+++ src/test/java/nextstep/waiting/ReservationWaitingE2ETest.java
@@ -1,7 +1,7 @@
 package nextstep.waiting;
 
-import auth.TokenRequest;
-import auth.TokenResponse;
+import auth.dto.TokenRequest;
+import auth.dto.TokenResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
